Fix averaging problem with IMU's
Testing field centric, at some points the IMU's would clash with their avgs
one being 0.01, the other being 1.99
This caused the problewm, as we would consider that to become 2 or 0,
but actually becomes 1. This (as I believe) caused the failure where,
two motors on opposite corners of the robot would spin in opposite directions,
causeing negligable movement. This is an untested fix.

Signed-off-by: Brozone0 <alexyou2006@outlook.com>